Create PostCard.tsx and refactor post lists
- Create a new `PostCard.tsx` component under `components/elements`.
- Move the post item logic from the top page's New Post section to this new component.
- Use the `PostCard.tsx` component for blog post links on the posts page and series pages.
- The new component has a `showMetadata` prop to control the display of extra details like categories and dates.
