+++ src/components/sections/NewPosts.tsx
@@ -1,5 +1,5 @@
 import { getAllPostTypes } from "@/lib/markdown";
-import PostItem from "@/components/elements/PostItem";
+import PostCard from "@/components/elements/PostCard";
 import Link from "next/link";
 
 const NewPosts = () => {
@@ -18,10 +18,11 @@ const NewPosts = () => {
       {/* 記事一覧 */}
       <div className="flex flex-col gap-16 md:gap-20 mb-12">
         {posts.slice(0, 3).map((post, index) => (
-          <PostItem
+          <PostCard
             key={post.slug}
             post={post}
-            isReversed={index % 2 !== 0} // 偶数番目と奇数番目でレイアウトを反転
+            isReversed={index % 2 !== 0}
+            showMetadata={false}
           />
         ))}
       </div>
